Add check to see if HDF5 volume input data is empty.
The HDF5 volume source processor crashes if it gets volume data it
cannot parse as its first input data.

This commit applies a bandaid by checking if the volume data list is
empty and doing nothing if that is the case. There are still some
fundamental issues that needs solving. The bandaid fix only workes
because the process function later throws the rank error exception due
to the selection setting being unitialised to a good value. If it were
to continue, the volume_ pointer will not have a valid value and
probably cause a crash.